+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *   Added ability to set zoom_factor (Dmytro Budnyk)
 *   Write JSON to the logger, rather than Ruby [Issue #430]
 *   Added ability to access all of a nodes attributes (Jon Rowe)
+*   Capybara 2.3 window support (Dmitry Vorotilin)
 *   Added ability to manipulate modals (`window.alert()`, `window.confirm()` and `window.prompt()`) that has been implemented from Capybara 2.4
 
 #### Bug fixes ####
@@ -16,6 +17,12 @@
 *   Render full window size when document has no height (Kevin McConnell)
 *   Don't alter focus with send_keys if we're already in the target element
     (Adam Prescott) [Issue #493]
+*   Close dup'ed fd after using (Dmitry Vorotilin) [Issue #446, #529, #528]
+
+### 1.5.1 ###
+
+#### Bug fixes ####
+
 *   Ensure process exits with correct status code and $! is propagated.
     (Micah Geisel) [Issue #497]